requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0

# Embeddings
sentence-transformers==3.0.0
torch>=2.0.0

# Vector database
faiss-cpu>=1.9.0
numpy>=1.24.0

# LLM communication
requests>=2.31.0

# Utilities
python-multipart==0.0.9
